requirements.txt
# Django y REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Servidor de producción
gunicorn==21.2.0
whitenoise==6.6.0

# Utilidades
python-dotenv==1.0.0
